refactor: common package 구조 변경
common package 구조 변경함
노션으로 먼저 구조를 정의한 뒤 변경
