+++ src/main/java/io/ratel/managesystem/common/http/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package io.ratel.managesystem.common.exception;
+package io.ratel.managesystem.common.http.response;
 
 public record ErrorResponse(
         String code,
