+++ src/main/java/io/ratel/managesystem/common/persistence/config/orm/JpaConfig.java
@@ -1,4 +1,4 @@
-package io.ratel.managesystem.common.persistence;
+package io.ratel.managesystem.common.persistence.config.orm;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
