+++ src/main/java/io/ratel/managesystem/domain/bookinfo/model/BookInfo.java
@@ -1,8 +1,7 @@
 package io.ratel.managesystem.domain.bookinfo.model;
 
-import io.ratel.managesystem.common.entity.AuditableFields;
+import io.ratel.managesystem.common.persistence.entity.AuditableFields;
 import io.ratel.managesystem.domain.bookinfo.entity.BookInfoEntity;
-import io.ratel.managesystem.domain.bookinfo.model.Category;
 import lombok.Getter;
 import lombok.Setter;
